backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
weaviate-client==4.9.6
python-dotenv==1.0.1
pydantic==2.10.4
langchain==0.3.14
langchain-groq==0.2.4


